+++ src/components/main/HotLectures.tsx
@@ -109,7 +109,7 @@ export default function HotLectures() {
                       variant="solid"
                       colorScheme="blue"
                       className="p-2 text-lightNavy hover:text-hoverNavy"
-                      onClick={() => { navigate(`/lecture/detail/${curriculum.curriculum_id}`)}}
+                      onClick={() => { navigate(`/curricula/detail/${curriculum.curriculum_id}`)}}
                     >
                       자세히 보기
                     </Button>
