+++ src/components/main/navBar/NavbarStudent.tsx
@@ -33,7 +33,7 @@ const NavbarStudent: React.FC<Props> = ({ username }) => {
   };
 
   return (
-    <nav className="sticky top-0 flex flex-wrap items-center justify-between p-2 bg-Beige border-b-2 border-hardBeige">
+    <nav className="sticky top-0 flex flex-wrap items-center justify-between p-2 bg-Beige border-b-2 border-hardBeige z-10">
       {/* Navbar 로고 */}
       <Link to={"/"}>
         <div className="w-28 ml-4">
